requirements.txt
psycopg2==2.9.10
python-dotenv==1.2.1
schedule==1.2.0
fastapi==0.115.0
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.7.0
pydantic-settings==2.2.0